+++ piccolo/query/methods/count.py
@@ -19,9 +19,14 @@ class Count(Query):
         super().__init__(table, **kwargs)
         self.where_delegate = WhereDelegate()
 
-    def where(self, *where: Combinable) -> Count:
+    ###########################################################################
+    # Clauses
+
+    def where(self: Self, *where: Combinable) -> Self:
         self.where_delegate.where(*where)
         return self
+
+    ###########################################################################
 
     async def response_handler(self, response) -> bool:
         return response[0]["count"]
@@ -36,3 +41,6 @@ class Count(Query):
                 select.querystrings[0],
             )
         ]
+
+
+Self = t.TypeVar("Self", bound=Count)